Opprydding + flere tester
