+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vedtak/læremidler/domain/Vedtaksperiode.kt
@@ -1,0 +1,13 @@
+package no.nav.tilleggsstonader.sak.vedtak.læremidler.domain
+
+import no.nav.tilleggsstonader.kontrakter.felles.Periode
+import java.time.LocalDate
+
+data class Vedtaksperiode(
+    override val fom: LocalDate,
+    override val tom: LocalDate,
+) : Periode<LocalDate> {
+    init {
+        validatePeriode()
+    }
+}
